refactor(injection): split injection.rs into per-concern submodules
language/injection.rs bundled five separable concerns in one flat file,
while the analogous analysis/semantic/ and analysis/selection/ modules
already decompose their multi-concept logic into per-concern files. This
extends that convention to the injection surface.

The free functions split cleanly by concern, with injection.rs kept as a
thin facade that re-exports the public surface (so every existing
crate::language::injection::X call site is unchanged):

- offset.rs    — #offset! directive parsing
- ranges.rs    — included-range math (compute/clip/sub-select/intersect)
- content.rs   — clean-content extraction + byte→Point coordinate helpers
- language.rs  — injection-language resolution for a matched pattern
- discovery.rs — region discovery, CacheableInjectionRegion, InjectionResolver

Cross-submodule helpers move to pub(super)/pub(crate) as the split
requires; test-only re-exports are cfg(test)-gated to keep the
production surface minimal. Pure structural Tidy: behavior and log
targets unchanged.

Refs #393
