+++ src/language/injection.rs
@@ -1,1263 +1,56 @@
-use crate::language::LanguageCoordinator;
-use crate::language::node_tracker::NodeTracker;
-use crate::language::predicate_accessor::{UnifiedPredicate, get_all_predicates};
-use crate::language::query_predicates::check_predicate;
-use crate::text::{floor_char_boundary, fnv1a_hash};
-use tree_sitter::{Node, Query, QueryCapture, QueryCursor, QueryMatch, StreamingIterator, Tree};
-use ulid::Ulid;
-use url::Url;
+//! Tree-sitter language injection: discovering injected regions (e.g. Lua in
+//! Markdown) and preparing them for highlighting and the LSP bridge.
+//!
+//! The surface is split by concern, mirroring `analysis/semantic/` and
+//! `analysis/selection/`:
+//! - [`offset`] — parsing the `#offset!` directive
+//! - [`ranges`] — computing the included byte ranges fed to the injection parser
+//! - [`content`] — clean-content extraction and byte→`Point` coordinate helpers
+//! - [`language`] — resolving the injection language for a matched pattern
+//! - [`discovery`] — finding regions and resolving them into cacheable form
+
+mod content;
+mod discovery;
+mod language;
+mod offset;
+mod ranges;
 
 /// Maximum recursion depth for nested injections to prevent stack overflow.
 pub(crate) const MAX_INJECTION_DEPTH: usize = 10;
 
-/// Represents offset adjustments for injection content boundaries
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Default)]
-pub struct InjectionOffset {
-    pub start_row: i32,
-    pub start_column: i32,
-    pub end_row: i32,
-    pub end_column: i32,
-}
-
-/// The `#offset!` directive for a pattern, normalized: a directive that
-/// parses to all zeros (malformed, or an explicit `#offset! … 0 0 0 0`) is a
-/// no-op and is reported as `None`, so consumers never disable
-/// included-range stripping — or skip raw-span fast paths — for an offset
-/// that changes nothing. Use this instead of
-/// [`parse_offset_directive_for_pattern`] anywhere behavior branches on the
-/// offset's presence.
-pub(crate) fn effective_offset_for_pattern(
-    query: &Query,
-    pattern_index: usize,
-) -> Option<InjectionOffset> {
-    parse_offset_directive_for_pattern(query, pattern_index)
-        .filter(|off| *off != InjectionOffset::default())
-}
-
-/// Parses offset directive for a specific pattern in the query.
-/// Returns None if the specified pattern has no #offset! directive for @injection.content.
-pub(crate) fn parse_offset_directive_for_pattern(
-    query: &Query,
-    pattern_index: usize,
-) -> Option<InjectionOffset> {
-    for predicate in get_all_predicates(query, pattern_index) {
-        // Skip non-offset! directives
-        if predicate.operator() != "offset!" {
-            continue;
-        }
-
-        // Skip non-General predicates
-        let UnifiedPredicate::General(pred) = predicate else {
-            continue;
-        };
-
-        // Skip if first arg is not a capture
-        let Some(tree_sitter::QueryPredicateArg::Capture(capture_id)) = pred.args.first() else {
-            continue;
-        };
-
-        // Skip if capture name not found or not @injection.content
-        let Some(_) = query
-            .capture_names()
-            .get(*capture_id as usize)
-            .filter(|name| **name == "injection.content")
-        else {
-            continue;
-        };
-
-        // Parse the 4 numeric arguments after the capture
-        // Format: (#offset! @injection.content start_row start_col end_row end_col)
-        let arg_count = pred.args.len();
-
-        // Validate argument count (should be 5: capture + 4 offsets)
-        if arg_count < 5 {
-            log::info!(
-                target: "kakehashi::query",
-                "Malformed #offset! directive for pattern {}: expected 4 offset values, got {}. \
-                Using default offset (0, 0, 0, 0). \
-                Correct format: (#offset! @injection.content start_row start_col end_row end_col)",
-                pattern_index,
-                arg_count - 1 // Subtract 1 for the capture argument
-            );
-            return Some(InjectionOffset::default());
-        }
-
-        // Try to parse each argument as i32
-        let parse_arg = |idx: usize| -> Result<i32, String> {
-            if let Some(tree_sitter::QueryPredicateArg::String(s)) = pred.args.get(idx) {
-                s.parse().map_err(|_| s.to_string())
-            } else {
-                Err(String::from("missing"))
-            }
-        };
-
-        // Parse all 4 offset values
-        let parse_results = [
-            ("start_row", parse_arg(1)),
-            ("start_col", parse_arg(2)),
-            ("end_row", parse_arg(3)),
-            ("end_col", parse_arg(4)),
-        ];
-
-        // Pattern match on all 4 results - more idiomatic than all(is_ok) + unwrap()
-        if let [
-            ("start_row", Ok(start_row)),
-            ("start_col", Ok(start_col)),
-            ("end_row", Ok(end_row)),
-            ("end_col", Ok(end_col)),
-        ] = parse_results.as_slice()
-        {
-            return Some(InjectionOffset {
-                start_row: *start_row,
-                start_column: *start_col,
-                end_row: *end_row,
-                end_column: *end_col,
-            });
-        }
-
-        // Log which values failed to parse
-        let error_details: Vec<String> = parse_results
-            .into_iter()
-            .filter_map(|(name, result)| result.err().map(|val| format!("{} = '{}'", name, val)))
-            .collect();
-
-        log::info!(
-            target: "kakehashi::query",
-            "Failed to parse #offset! directive for pattern {}: invalid values [{}]. \
-            Using default offset (0, 0, 0, 0). \
-            All offset values must be integers.",
-            pattern_index,
-            error_details.join(", ")
-        );
-
-        return Some(InjectionOffset::default());
-    }
-    None
-}
-
-/// Iterates over `@injection.content` captures that pass general predicate filtering.
-///
-/// Combines predicate evaluation (e.g., `#lua-match?`) and capture name checking
-/// into a single iterator, eliminating duplication between `collect_all_injections`
-/// and `extract_content_and_language`.
-fn iter_valid_injection_content_captures<'a, 'b>(
-    match_: &'b QueryMatch<'_, 'a>,
-    query: &'b Query,
-    text: &'b str,
-) -> impl Iterator<Item = QueryCapture<'a>> + 'b {
-    match_.captures.iter().copied().filter(move |capture| {
-        if !check_predicate(query, match_, capture, text) {
-            return false;
-        }
-        query
-            .capture_names()
-            .get(capture.index as usize)
-            .is_some_and(|name| *name == "injection.content")
-    })
-}
-
-/// Checks if a node is within the bounds of another node
-fn is_node_within(node: &Node, container: &Node) -> bool {
-    node.start_byte() >= container.start_byte() && node.end_byte() <= container.end_byte()
-}
-
-/// Extracts the injection language from query properties or captures.
-///
-/// Tries static `#set!`, then nvim-treesitter's `#set-lang-from-info-string!`,
-/// then a dynamic `@injection.language` capture, in that order.
-fn extract_injection_language(query: &Query, match_: &QueryMatch, text: &str) -> Option<String> {
-    // First check for static language via #set! property
-    if let Some(language) = extract_static_language(query, match_) {
-        return Some(language);
-    }
-
-    // Then check for nvim-treesitter's #set-lang-from-info-string! predicate
-    if let Some(language) = extract_language_from_info_string(query, match_, text) {
-        return Some(language);
-    }
-
-    // Finally check for dynamic language via @injection.language capture
-    extract_dynamic_language(query, match_, text)
-}
-
-/// Checks whether the given pattern has `#set! injection.combined`.
-///
-/// Like `injection.include-children`, this is a value-less property — its
-/// **presence** means all `@injection.content` captures of the pattern (for
-/// the same language) are parsed as one document via `set_included_ranges`,
-/// so cross-region context survives (e.g. an HTML tag opened in one
-/// `html_block` and closed in another). See #187.
-pub(crate) fn has_combined_for_pattern(query: &Query, pattern_index: usize) -> bool {
-    for predicate in get_all_predicates(query, pattern_index) {
-        if let UnifiedPredicate::Property(prop) = predicate
-            && prop.key.as_ref() == "injection.combined"
-        {
-            return true;
-        }
-    }
-    false
-}
-
-/// Checks whether the given pattern has `#set! injection.include-children`.
-///
-/// This is a value-less property — its **presence** alone means "include children"
-/// (the injection parser sees the full content node including named children).
-/// When absent, named children should be excluded via `set_included_ranges()`.
-pub(crate) fn has_include_children_for_pattern(query: &Query, pattern_index: usize) -> bool {
-    for predicate in get_all_predicates(query, pattern_index) {
-        if let UnifiedPredicate::Property(prop) = predicate
-            && prop.key.as_ref() == "injection.include-children"
-        {
-            return true;
-        }
-    }
-    false
-}
-
-/// Compute the included ranges for an injection content node.
-///
-/// When `include_children` is false and the content node has named children
-/// (e.g., `block_continuation` nodes in blockquote code blocks), this function
-/// computes the "gap" ranges between those children — the actual code content.
-///
-/// Returns `None` when:
-/// - `include_children` is true (the injection parser should see everything)
-/// - The content node has no named children (no gaps to compute)
-/// - All content is covered by named children (no gaps exist)
-///
-/// The returned ranges are **relative** to the content node's start position:
-/// byte offsets are subtracted by `content_node.start_byte()`, and row/column
-/// Points are adjusted relative to `content_node.start_position()`.
-pub(crate) fn compute_included_ranges(
-    content_node: &tree_sitter::Node,
-    include_children: bool,
-) -> Option<Vec<tree_sitter::Range>> {
-    compute_included_ranges_in_window(
-        content_node,
-        include_children,
-        content_node.start_byte(),
-        content_node.end_byte(),
-        || (content_node.start_position(), content_node.end_position()),
-    )
-}
-
-/// Like [`compute_included_ranges`], but restricted to an effective byte
-/// window — the post-`#offset!` span (#186). Children outside the window are
-/// ignored; children straddling a window edge are clipped to it. Returned
-/// ranges are relative to `window.start`, matching the content slice handed
-/// to the injection parser.
-///
-/// `window` bounds must lie on char boundaries (callers snap with
-/// `ceil_char_boundary` / `floor_char_boundary` before slicing the content).
-pub(crate) fn compute_included_ranges_clipped(
-    content_node: &tree_sitter::Node,
-    include_children: bool,
-    text: &str,
-    window: std::ops::Range<usize>,
-) -> Option<Vec<tree_sitter::Range>> {
-    if window.start >= window.end {
-        return None;
-    }
-    compute_included_ranges_in_window(
-        content_node,
-        include_children,
-        window.start,
-        window.end,
-        || {
-            // Anchor the scans on the content node's cached position so we only
-            // walk the (typically small) span between the node start and the
-            // window edges, not the whole document prefix. Lazy: skipped entirely
-            // when the guards in the core return early.
-            let anchor_byte = content_node.start_byte();
-            let anchor_point = content_node.start_position();
-            let start = byte_to_point_anchored(text, window.start, anchor_byte, anchor_point);
-            let end = byte_to_point_anchored(text, window.end, window.start, start);
-            (start, end)
-        },
-    )
-}
-
-/// Shared core of [`compute_included_ranges`] /
-/// [`compute_included_ranges_clipped`]: gaps between the content node's named
-/// children, restricted to `[window_start_byte, window_end_byte)` and
-/// relativized to the window start. `window_points` lazily supplies the
-/// window bounds' Points — it is only invoked once the guards have decided
-/// gaps actually need computing.
-fn compute_included_ranges_in_window(
-    content_node: &tree_sitter::Node,
-    include_children: bool,
-    window_start_byte: usize,
-    window_end_byte: usize,
-    window_points: impl FnOnce() -> (tree_sitter::Point, tree_sitter::Point),
-) -> Option<Vec<tree_sitter::Range>> {
-    if include_children || content_node.named_child_count() == 0 {
-        return None;
-    }
-
-    // Check if all named children are zero-width (ghost nodes from
-    // tree-sitter parsing with included_ranges). If so, treat as if
-    // there are no children — the gap computation would produce a
-    // trivial single range covering the entire node.
-    let has_nonzero_children = {
-        let mut cursor = content_node.walk();
-        content_node
-            .named_children(&mut cursor)
-            .any(|c| c.start_byte() != c.end_byte())
-    };
-    if !has_nonzero_children {
-        return None;
-    }
-
-    let (window_start_point, window_end_point) = window_points();
-
-    // Helper to make a Point relative to the window's start.
-    // Column is only adjusted when the point is on the same row as the window
-    // start, because only that row has a non-zero column offset from the
-    // window origin.
-    let relativize_point = |p: tree_sitter::Point| -> tree_sitter::Point {
-        tree_sitter::Point {
-            row: p.row.saturating_sub(window_start_point.row),
-            column: if p.row == window_start_point.row {
-                p.column.saturating_sub(window_start_point.column)
-            } else {
-                p.column
-            },
-        }
-    };
-
-    let mut ranges = Vec::new();
-    let mut cursor_byte = window_start_byte;
-    let mut cursor_point = window_start_point;
-
-    let mut tree_cursor = content_node.walk();
-    for child in content_node.named_children(&mut tree_cursor) {
-        let child_start_byte = child.start_byte();
-        let child_end_byte = child.end_byte();
-
-        // Skip zero-width children. Tree-sitter-markdown creates zero-width
-        // block_continuation nodes when parsing with included_ranges that
-        // already excluded the `> ` prefix bytes. These ghost nodes don't
-        // span any bytes, so they don't define meaningful gaps.
-        if child_start_byte == child_end_byte {
-            continue;
-        }
-
-        // Entirely before the window (or before an earlier child that already
-        // advanced the cursor): nothing to exclude here.
-        if child_end_byte <= cursor_byte {
-            continue;
-        }
-        // Children are in document order; the first one at/after the window
-        // end means no later child can affect the window either.
-        if child_start_byte >= window_end_byte {
-            break;
-        }
-
-        // Gap before this child
-        if cursor_byte < child_start_byte {
-            ranges.push(tree_sitter::Range {
-                start_byte: cursor_byte - window_start_byte,
-                end_byte: child_start_byte - window_start_byte,
-                start_point: relativize_point(cursor_point),
-                end_point: relativize_point(child.start_position()),
-            });
-        }
-
-        // A child running past the window end is clipped: the window's tail
-        // is covered, so no final gap remains.
-        if child_end_byte >= window_end_byte {
-            cursor_byte = window_end_byte;
-            cursor_point = window_end_point;
-            break;
-        }
-
-        cursor_byte = child_end_byte;
-        cursor_point = child.end_position();
-    }
-
-    // Gap after last child
-    if cursor_byte < window_end_byte {
-        ranges.push(tree_sitter::Range {
-            start_byte: cursor_byte - window_start_byte,
-            end_byte: window_end_byte - window_start_byte,
-            start_point: relativize_point(cursor_point),
-            end_point: relativize_point(window_end_point),
-        });
-    }
-
-    if ranges.is_empty() {
-        None
-    } else {
-        Some(ranges)
-    }
-}
-
-/// Clip the parent's `included_ranges` to the nested injection's byte window
-/// and re-relativize. Used when the nested injection's tree was parsed with
-/// `included_ranges` and has no `block_continuation` children of its own;
-/// both byte arguments are relative to `parent_ranges`' base (parent
-/// `content_text` start). Returns `None` if no parent range overlaps.
-///
-/// Known limitation: when a range is clipped by `max`/`min`, only its bytes
-/// are adjusted — `start_point`/`end_point` are not. Tree-sitter accepts this
-/// (it validates byte ordering only), but nodes from clipped ranges may
-/// report slightly wrong columns. Rare in practice: nested boundaries
-/// usually align with parent gap boundaries.
-pub(crate) fn sub_select_included_ranges(
-    parent_ranges: &[tree_sitter::Range],
-    nested_start_byte: usize,
-    nested_end_byte: usize,
-) -> Option<Vec<tree_sitter::Range>> {
-    // Find the base row from the first overlapping range
-    let first_overlap = parent_ranges
-        .iter()
-        .find(|r| r.end_byte > nested_start_byte && r.start_byte < nested_end_byte)?;
-    let base_row = first_overlap.start_point.row;
-
-    let mut clipped = Vec::new();
-    let mut is_first = true;
-
-    for r in parent_ranges {
-        // Skip non-overlapping ranges
-        if r.end_byte <= nested_start_byte || r.start_byte >= nested_end_byte {
-            continue;
-        }
-
-        // Clip to nested region
-        let clip_start = r.start_byte.max(nested_start_byte);
-        let clip_end = r.end_byte.min(nested_end_byte);
-
-        // The first sub-selected range's column must be 0 because the prefix
-        // bytes (e.g., "> ") preceding this range are NOT in the nested
-        // content_text — they come before nested_start_byte. Without this,
-        // the column would double-count with content_start_col in the token
-        // collector. Subsequent ranges keep the parent's column because their
-        // prefix bytes ARE within the nested content_text.
-        let start_column = if is_first { 0 } else { r.start_point.column };
-
-        clipped.push(tree_sitter::Range {
-            start_byte: clip_start.saturating_sub(nested_start_byte),
-            end_byte: clip_end.saturating_sub(nested_start_byte),
-            start_point: tree_sitter::Point {
-                row: r.start_point.row.saturating_sub(base_row),
-                column: start_column,
-            },
-            end_point: tree_sitter::Point {
-                row: r.end_point.row.saturating_sub(base_row),
-                column: r.end_point.column,
-            },
-        });
-
-        is_first = false;
-    }
-
-    if clipped.is_empty() {
-        None
-    } else {
-        Some(clipped)
-    }
-}
-
-/// Intersect two sets of included ranges, keeping only byte regions present in both.
-///
-/// Both range sets must be in the same coordinate space (e.g., both relative
-/// to `content_node.start_byte()`). The result contains only byte regions
-/// that appear in BOTH inputs.
-///
-/// This is needed when a tree parsed with parent `included_ranges` produces
-/// `compute_included_ranges()` gap ranges that inadvertently span parent-excluded
-/// bytes. Intersecting with the parent's `sub_select_included_ranges()` output
-/// ensures previously-excluded bytes stay excluded.
-pub(crate) fn intersect_included_ranges(
-    a: &[tree_sitter::Range],
-    b: &[tree_sitter::Range],
-) -> Vec<tree_sitter::Range> {
-    let mut result = Vec::new();
-    let mut j = 0;
-
-    for ar in a {
-        // Advance b past ranges that end before ar starts
-        while j < b.len() && b[j].end_byte <= ar.start_byte {
-            j += 1;
-        }
-
-        // Check all b ranges that overlap with ar
-        let mut k = j;
-        while k < b.len() && b[k].start_byte < ar.end_byte {
-            let start_byte = ar.start_byte.max(b[k].start_byte);
-            let end_byte = ar.end_byte.min(b[k].end_byte);
-
-            if start_byte < end_byte {
-                // Use point from whichever boundary is more restrictive
-                let start_point = if start_byte == ar.start_byte {
-                    ar.start_point
-                } else {
-                    b[k].start_point
-                };
-                let end_point = if end_byte == ar.end_byte {
-                    ar.end_point
-                } else {
-                    b[k].end_point
-                };
-
-                result.push(tree_sitter::Range {
-                    start_byte,
-                    end_byte,
-                    start_point,
-                    end_point,
-                });
-            }
-            k += 1;
-        }
-    }
-
-    result
-}
-
-/// Extract clean injection content, stripping child node bytes when `included_ranges` present.
-///
-/// When `included_ranges` is `Some`, only the gap bytes (actual code content) are
-/// concatenated, effectively stripping blockquote prefixes like `> `. When `None`,
-/// returns the full content slice as-is.
-pub(crate) fn extract_clean_content(
-    host_text: &str,
-    byte_range: std::ops::Range<usize>,
-    included_ranges: Option<&[tree_sitter::Range]>,
-) -> String {
-    let content = &host_text[byte_range.clone()];
-    match included_ranges {
-        None => content.to_string(),
-        Some(ranges) => {
-            let capacity: usize = ranges.iter().map(|r| r.end_byte - r.start_byte).sum();
-            let mut result = String::with_capacity(capacity);
-            for range in ranges {
-                result.push_str(&content[range.start_byte..range.end_byte]);
-            }
-            result
-        }
-    }
-}
-
-/// Compute per-virtual-line UTF-16 column offsets from included ranges.
-///
-/// Each offset is the width of the blockquote prefix (or other excluded content)
-/// before the gap on that line. When `included_ranges` is `None`, returns
-/// `vec![start_column]` — the non-blockquote fallback where only line 0 has an offset.
-pub(crate) fn compute_line_column_offsets(
-    host_text: &str,
-    byte_range: std::ops::Range<usize>,
-    start_column: u32,
-    included_ranges: Option<&[tree_sitter::Range]>,
-) -> Vec<u32> {
-    match included_ranges {
-        None => vec![start_column],
-        Some(ranges) => {
-            // Build a map from virtual line → column offset.
-            // Each gap range starts at a known point with a column value.
-            // The column value represents how many columns of prefix exist
-            // before the actual code on that line.
-            let mut offsets = Vec::new();
-            for range in ranges {
-                let gap_start_line = range.start_point.row;
-                // How many virtual lines does this gap span?
-                let gap_end_line = range.end_point.row;
-                for line in gap_start_line..=gap_end_line {
-                    offsets.resize(line + 1, 0);
-                    if line == gap_start_line {
-                        // The column offset is the gap's start column within the host line.
-                        // For the first line (row 0) of the content node, use start_column
-                        // (which accounts for the full prefix from the line start).
-                        // For subsequent lines, the start_point.column from
-                        // compute_included_ranges is already the raw byte column
-                        // within the line — convert to UTF-16.
-                        if line == 0 {
-                            offsets[line] = start_column;
-                        } else {
-                            // Convert byte column to UTF-16 code units.
-                            // The gap's start_byte is relative to content node start.
-                            // The line starts at the byte after the previous newline.
-                            let gap_abs_byte = byte_range.start + range.start_byte;
-                            let line_start_byte = gap_abs_byte - range.start_point.column;
-                            let prefix = &host_text[line_start_byte..gap_abs_byte];
-                            offsets[line] = prefix.encode_utf16().count() as u32;
-                        }
-                    }
-                }
-            }
-            if offsets.is_empty() {
-                vec![start_column]
-            } else {
-                offsets
-            }
-        }
-    }
-}
-
-/// Compute clean virtual content and per-line column offsets for an injection region.
-///
-/// This combines `compute_included_ranges`, `extract_clean_content`, and
-/// `compute_line_column_offsets` into a single call, avoiding duplication
-/// across resolve methods.
-fn extract_virtual_content_and_offsets(
-    region: &InjectionRegionInfo,
-    cacheable: &CacheableInjectionRegion,
-    text: &str,
-) -> (String, Vec<u32>) {
-    // Child-exclusion gaps restricted to the effective window (#186):
-    // cacheable.byte_range already reflects any #offset! directive (applied
-    // and char-boundary-aligned by from_region_info), so gaps are clipped to
-    // it and relativized to its start. Without an offset the window equals
-    // the content node span, so the node-anchored variant gives the same
-    // result while reusing tree-sitter's cached node Points (no text scan).
-    let included_ranges = if region.offset.is_some() {
-        compute_included_ranges_clipped(
-            &region.content_node,
-            region.include_children,
-            text,
-            cacheable.byte_range.clone(),
-        )
-    } else {
-        compute_included_ranges(&region.content_node, region.include_children)
-    };
-    let virtual_content = extract_clean_content(
-        text,
-        cacheable.byte_range.clone(),
-        included_ranges.as_deref(),
-    );
-    let line_column_offsets = compute_line_column_offsets(
-        text,
-        cacheable.byte_range.clone(),
-        cacheable.start_column,
-        included_ranges.as_deref(),
-    );
-    (virtual_content, line_column_offsets)
-}
-
-/// Parse text with optional included ranges, resetting parser state afterward.
-///
-/// Shared protocol for Tree-sitter's `set_included_ranges` API; both semantic
-/// tokens and selection range paths delegate here to avoid divergence. The reset
-/// to `&[]` is unconditional (even on the failure path) so parsers returned to
-/// pools or caches never carry stale included-range state.
-pub(crate) fn parse_with_ranges(
-    parser: &mut tree_sitter::Parser,
-    text: &str,
-    included_ranges: Option<&[tree_sitter::Range]>,
-    log_target: &str,
-    lang_name: &str,
-) -> Option<tree_sitter::Tree> {
-    if let Some(ranges) = included_ranges
-        && let Err(e) = parser.set_included_ranges(ranges)
-    {
-        log::warn!(
-            target: log_target,
-            "Failed to set included ranges for {}: {}. Skipping parse.",
-            lang_name, e
-        );
-        let _ = parser.set_included_ranges(&[]);
-        return None;
-    }
-
-    let tree = parser.parse(text, None);
-
-    // Always reset included ranges after parsing to prevent stale state
-    let _ = parser.set_included_ranges(&[]);
-
-    tree
-}
-
-/// Extracts language from #set! injection.language property
-fn extract_static_language(query: &Query, match_: &QueryMatch) -> Option<String> {
-    // Use unified accessor to check property settings
-    for predicate in get_all_predicates(query, match_.pattern_index) {
-        if let UnifiedPredicate::Property(prop) = predicate
-            && prop.key.as_ref() == "injection.language"
-            && let Some(value) = &prop.value
-        {
-            return Some(value.as_ref().to_string());
-        }
-    }
-    None
-}
-
-/// Extracts language from @injection.language capture
-fn extract_dynamic_language(query: &Query, match_: &QueryMatch, text: &str) -> Option<String> {
-    for capture in match_.captures {
-        if let Some(capture_name) = query.capture_names().get(capture.index as usize)
-            && *capture_name == "injection.language"
-        {
-            let lang_text = &text[capture.node.byte_range()];
-            return Some(lang_text.to_string());
-        }
-    }
-    None
-}
-
-/// Extracts language from nvim-treesitter's `#set-lang-from-info-string!` predicate.
-///
-/// This custom predicate uses a capture's text as the injection language, commonly
-/// for markdown fenced code blocks.
-fn extract_language_from_info_string(
-    query: &Query,
-    match_: &QueryMatch,
-    text: &str,
-) -> Option<String> {
-    // Look for #set-lang-from-info-string! predicate
-    for predicate in get_all_predicates(query, match_.pattern_index) {
-        if predicate.operator() == "set-lang-from-info-string!"
-            && let UnifiedPredicate::General(pred) = predicate
-        {
-            // The predicate takes a capture reference as argument
-            if let Some(tree_sitter::QueryPredicateArg::Capture(capture_id)) = pred.args.first() {
-                // Find the capture in the match
-                for capture in match_.captures {
-                    if capture.index == *capture_id {
-                        // Extract the text from the captured node as the language
-                        let lang_text = &text[capture.node.byte_range()];
-                        // Normalize the language name (lowercase, trim)
-                        let normalized = lang_text.trim().to_lowercase();
-                        if !normalized.is_empty() {
-                            return Some(normalized);
-                        }
-                    }
-                }
-            }
-        }
-    }
-    None
-}
-
-/// Represents an injection region found in the document
-#[derive(Debug, Clone)]
-pub struct InjectionRegionInfo<'a> {
-    /// The injection language (e.g., "lua", "yaml")
-    pub language: String,
-    /// The content node from the injection query
-    pub content_node: Node<'a>,
-    /// The pattern index (for offset directive lookups)
-    pub pattern_index: usize,
-    /// Whether `#set! injection.include-children` is set for this pattern.
-    /// When true, the injection parser sees the full content node (including named children).
-    /// When false, named children (e.g., `block_continuation`) should be excluded.
-    pub include_children: bool,
-    /// The `#offset!` directive for this pattern, resolved at collection time
-    /// (the query goes out of scope before consumers like
-    /// `CacheableInjectionRegion::from_region_info` run). `None` when the
-    /// pattern has no directive.
-    pub offset: Option<InjectionOffset>,
-}
-
-use std::ops::Range;
-
-/// Owned injection region for caching (no lifetime dependency on parse tree)
-///
-/// Unlike `InjectionRegionInfo<'a>`, this struct owns all its data and can be
-/// stored in caches that outlive the parse tree. Created via `from_region_info()`.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct CacheableInjectionRegion {
-    /// The injection language (e.g., "lua", "yaml")
-    pub language: String,
-    /// Byte range of the injection content in the source
-    pub byte_range: Range<usize>,
-    /// Line range (0-indexed, inclusive start, exclusive end)
-    pub line_range: Range<u32>,
-    /// Column offset of the injection content start within its start line,
-    /// stored as UTF-16 code units (matching LSP's position encoding).
-    /// Non-zero for inline injections (e.g., code within a markdown paragraph).
-    /// Only affects the first line (line 0) of the virtual document during
-    /// coordinate translation; subsequent lines start at column 0.
-    pub start_column: u32,
-    /// Unique identifier for associating with cached tokens.
-    /// Generated by NodeTracker as a position-based ULID.
-    pub region_id: String,
-    /// Hash of the injection content for cache invalidation.
-    /// Used to detect when cached semantic tokens should be invalidated:
-    /// if content_hash changes (or language changes), the cached tokens are stale.
-    pub content_hash: u64,
-}
-
-/// Compute the (row, UTF-16 column) of `byte_pos`, using a nearby byte with a
-/// known row as an anchor: the row counts newlines only in the span between
-/// them (offset deltas are typically a few lines, never the whole document).
-/// The column additionally scans back from `byte_pos` to its line start —
-/// bounded by the current line's length, or by `byte_pos` itself in a
-/// document without newlines.
-fn position_of_byte(
-    text: &str,
-    byte_pos: usize,
-    anchor_byte: usize,
-    anchor_row: usize,
-) -> (u32, u32) {
-    let (lo, hi) = (byte_pos.min(anchor_byte), byte_pos.max(anchor_byte));
-    let newlines = text[lo..hi].bytes().filter(|&b| b == b'\n').count();
-    let row = if byte_pos >= anchor_byte {
-        anchor_row + newlines
-    } else {
-        // Defensive: a stale tree whose row metadata disagrees with `text`
-        // must not panic the server on underflow.
-        anchor_row.saturating_sub(newlines)
-    };
-    let line_start = text[..byte_pos].rfind('\n').map(|i| i + 1).unwrap_or(0);
-    let column = text[line_start..byte_pos].encode_utf16().count() as u32;
-    (row as u32, column)
-}
-
-/// Convert an absolute byte offset to a `tree_sitter::Point` (byte-based
-/// column). Used when an offset directive shifts the injection boundary away
-/// from a known node position, so we can't reuse the content node's
-/// start/end points.
-pub(crate) fn byte_to_point(text: &str, byte: usize) -> tree_sitter::Point {
-    // Align first — slicing `&text[..clamped]` on a mid-character byte would
-    // panic and crash the LSP server.
-    let clamped = floor_char_boundary(text, byte);
-    let prefix = &text[..clamped];
-    let row = prefix.bytes().filter(|b| *b == b'\n').count();
-    let last_nl = prefix.rfind('\n');
-    let column = match last_nl {
-        Some(idx) => clamped - idx - 1,
-        None => clamped,
-    };
-    tree_sitter::Point { row, column }
-}
-
-/// Like [`byte_to_point`], but scans only the span between a known anchor
-/// and `byte` instead of the whole prefix of `text` — the anchor is
-/// typically a tree-sitter node's cached `start_position()`. Falls back to a
-/// full scan when `byte` precedes the anchor (e.g. a negative `#offset!`
-/// extending before the content node). `anchor_byte` must lie on a char
-/// boundary, with `anchor_point` its Point in `text`.
-pub(crate) fn byte_to_point_anchored(
-    text: &str,
-    byte: usize,
-    anchor_byte: usize,
-    anchor_point: tree_sitter::Point,
-) -> tree_sitter::Point {
-    let clamped = floor_char_boundary(text, byte);
-    if clamped < anchor_byte {
-        return byte_to_point(text, clamped);
-    }
-    let segment = &text[anchor_byte..clamped];
-    match segment.rfind('\n') {
-        Some(last_nl) => tree_sitter::Point {
-            row: anchor_point.row + segment.bytes().filter(|b| *b == b'\n').count(),
-            column: segment.len() - last_nl - 1,
-        },
-        None => tree_sitter::Point {
-            row: anchor_point.row,
-            column: anchor_point.column + segment.len(),
-        },
-    }
-}
-
-impl CacheableInjectionRegion {
-    /// Create from an InjectionRegionInfo, extracting position data from the node.
-    ///
-    /// # Panics (debug builds)
-    /// Panics if the content node is zero-width (`start_byte >= end_byte`).
-    /// Callers must filter zero-width nodes upstream (see `collect_all_injections`).
-    pub fn from_region_info(info: &InjectionRegionInfo<'_>, region_id: &str, text: &str) -> Self {
-        let node = &info.content_node;
-        debug_assert!(
-            node.start_byte() < node.end_byte(),
-            "from_region_info called with zero-width node at byte {}",
-            node.start_byte(),
-        );
-
-        // #offset! narrows the raw node span to the effective content range
-        // (e.g. trimming `---` frontmatter delimiters). The bridge consumes
-        // byte_range / line_range / start_column for virtual-document content
-        // extraction and coordinate translation, so all of them must reflect
-        // the effective range, mirroring the semantic path (#183).
-        let (start_byte, end_byte) = match info.offset {
-            Some(offset) => {
-                use crate::analysis::offset_calculator::{ByteRange, calculate_effective_range};
-                // calculate_effective_range clamps, snaps inward to char
-                // boundaries, and normalizes start <= end, so slicing below
-                // cannot panic.
-                let effective = calculate_effective_range(
-                    text,
-                    ByteRange::new(node.start_byte(), node.end_byte()),
-                    offset,
-                );
-                (effective.start, effective.end)
-            }
-            None => (node.start_byte(), node.end_byte()),
-        };
-
-        let content = &text[start_byte..end_byte];
-
-        // Convert tree-sitter byte column to UTF-16 code units for LSP compatibility.
-        // Tree-sitter reports columns as byte offsets, but LSP positions use UTF-16.
-        // For ASCII they're identical, but non-ASCII prefixes (CJK, emoji) differ.
-        // tree-sitter's `column` is a byte offset from the line start, so
-        // `start_byte() - column` correctly recovers the line-start byte even
-        // when the line contains multi-byte UTF-8 characters before the node.
-        let (start_line, start_column) = if start_byte == node.start_byte() {
-            let byte_column = node.start_position().column;
-            let line_start_byte = node.start_byte() - byte_column;
-            let line_prefix = &text[line_start_byte..node.start_byte()];
-            (
-                node.start_position().row as u32,
-                line_prefix.encode_utf16().count() as u32,
-            )
-        } else {
-            position_of_byte(
-                text,
-                start_byte,
-                node.start_byte(),
-                node.start_position().row,
-            )
-        };
-
-        // end_position() points past the last byte. When that position is mid-line
-        // (column > 0), the node still occupies that row → exclusive end is row + 1.
-        // When column == 0 (node ended with newline), the row is already one past
-        // the last occupied line — use it directly.
-        let end_line = if end_byte == node.end_byte() {
-            let end_pos = node.end_position();
-            if end_pos.column > 0 {
-                end_pos.row as u32 + 1
-            } else {
-                end_pos.row as u32
-            }
-        } else {
-            let (end_row, end_column) =
-                position_of_byte(text, end_byte, node.end_byte(), node.end_position().row);
-            if end_column > 0 { end_row + 1 } else { end_row }
-        };
-
-        Self {
-            language: info.language.clone(),
-            byte_range: start_byte..end_byte,
-            line_range: start_line..end_line,
-            start_column,
-            region_id: region_id.to_string(),
-            content_hash: Self::hash_content(content),
-        }
-    }
-
-    /// Compute a simple hash of content bytes for stable matching.
-    fn hash_content(content: &str) -> u64 {
-        fnv1a_hash(content)
-    }
-}
-
-/// Collects all injection regions in the document.
-///
-/// Unlike `detect_injection` (which requires a specific node), this finds ALL
-/// injection regions in the whole document — used by semantic tokens to highlight
-/// every injected region. `None` when there is no query.
-pub fn collect_all_injections<'a>(
-    root: &Node<'a>,
-    text: &str,
-    injection_query: Option<&Query>,
-) -> Option<Vec<InjectionRegionInfo<'a>>> {
-    let query = injection_query?;
-
-    let mut cursor = QueryCursor::new();
-    let mut matches = cursor.matches(query, *root, text.as_bytes());
-
-    // Use a map to deduplicate by content node range
-    let mut injections_map = std::collections::HashMap::new();
-
-    while let Some(match_) = matches.next() {
-        for capture in iter_valid_injection_content_captures(match_, query, text) {
-            if capture.node.start_byte() >= capture.node.end_byte() {
-                continue;
-            }
-            if let Some(language) = extract_injection_language(query, match_, text) {
-                let key = (capture.node.start_byte(), capture.node.end_byte());
-                injections_map.entry(key).or_insert(InjectionRegionInfo {
-                    language,
-                    content_node: capture.node,
-                    pattern_index: match_.pattern_index,
-                    include_children: has_include_children_for_pattern(query, match_.pattern_index),
-                    offset: effective_offset_for_pattern(query, match_.pattern_index),
-                });
-            }
-        }
-    }
-
-    // Sort by start_byte (primary) and end_byte (secondary) to ensure deterministic ordering
-    let mut injections: Vec<_> = injections_map.into_values().collect();
-    injections.sort_by_key(|r| (r.content_node.start_byte(), r.content_node.end_byte()));
-    Some(injections)
-}
-
-/// Detects injection and returns both the language and the content node
-/// Also returns the pattern index of the innermost injection for offset lookups
-pub(crate) fn detect_injection<'a>(
-    node: &Node<'a>,
-    root: &Node<'a>,
-    text: &str,
-    injection_query: Option<&Query>,
-    base_language: &str,
-) -> Option<(Vec<String>, Node<'a>, usize)> {
-    let injections = collect_injection_regions(node, root, text, injection_query)?;
-
-    if injections.is_empty() {
-        return None;
-    }
-
-    // Sort injections by their range (outer to inner)
-    let mut sorted_injections = injections;
-    sorted_injections.sort_by(|a, b| {
-        // Sort by start byte (ascending), then by end byte (descending)
-        // This ensures outer injections come before inner ones
-        a.start_byte
-            .cmp(&b.start_byte)
-            .then(b.end_byte.cmp(&a.end_byte))
-    });
-
-    // Build the language hierarchy from outermost to innermost
-    let mut hierarchy = vec![base_language.to_string()];
-    for region in &sorted_injections {
-        hierarchy.push(region.language.clone());
-    }
-
-    // Return the innermost content node and its pattern index
-    let innermost = sorted_injections.last()?;
-
-    Some((hierarchy, innermost.content_node, innermost.pattern_index))
-}
-
-/// Represents an injection region with its metadata (unresolved intermediate data).
-struct RawInjectionRegion<'a> {
-    start_byte: usize,
-    end_byte: usize,
-    language: String,
-    content_node: Node<'a>,
-    pattern_index: usize,
-}
-
-/// Collects all injection regions that contain the given node
-/// Returns a list of `RawInjectionRegion` values for each injection containing the node.
-fn collect_injection_regions<'a>(
-    node: &Node<'a>,
-    root: &Node<'a>,
-    text: &str,
-    injection_query: Option<&Query>,
-) -> Option<Vec<RawInjectionRegion<'a>>> {
-    let query = injection_query?;
-
-    // Run the query on the entire tree
-    let mut cursor = QueryCursor::new();
-    let mut matches = cursor.matches(query, *root, text.as_bytes());
-
-    // Collect all injection regions that contain our node
-    // Use a map to deduplicate by node range (start, end)
-    let mut injections_map = std::collections::HashMap::new();
-
-    while let Some(match_) = matches.next() {
-        if let Some((content_node, language, pattern_index)) =
-            extract_content_and_language(node, match_, query, text)
-        {
-            let key = (content_node.start_byte(), content_node.end_byte());
-
-            // Only keep the first injection for each unique range
-            // This handles cases where multiple patterns match the same node
-            injections_map.entry(key).or_insert(RawInjectionRegion {
-                start_byte: content_node.start_byte(),
-                end_byte: content_node.end_byte(),
-                language,
-                content_node,
-                pattern_index,
-            });
-        }
-    }
-
-    // Convert to vector
-    let injections: Vec<_> = injections_map.into_values().collect();
-
-    Some(injections)
-}
-
-/// Extracts the injection content node and language if the given node is within it
-/// Also returns the pattern index for offset lookups
-fn extract_content_and_language<'a>(
-    node: &Node<'a>,
-    match_: &QueryMatch<'_, 'a>,
-    query: &Query,
-    text: &str,
-) -> Option<(Node<'a>, String, usize)> {
-    for capture in iter_valid_injection_content_captures(match_, query, text) {
-        let content_node = capture.node;
-
-        // Check if our node is within this injection region
-        if is_node_within(node, &content_node) {
-            // Extract the injection language
-            if let Some(language) = extract_injection_language(query, match_, text) {
-                return Some((content_node, language, match_.pattern_index));
-            }
-        }
-    }
-
-    None
-}
-
-/// Find the injection region containing the given byte offset.
-///
-/// Returns `(index, region)` for use with `calculate_region_id`, or `None` when the
-/// position is not within any injection region.
-pub(crate) fn find_injection_at_position<'a>(
-    injections: &'a [InjectionRegionInfo<'a>],
-    byte_offset: usize,
-) -> Option<(usize, &'a InjectionRegionInfo<'a>)> {
-    injections.iter().enumerate().find(|(_, inj)| {
-        let start = inj.content_node.start_byte();
-        let end = inj.content_node.end_byte();
-        byte_offset >= start && byte_offset < end
-    })
-}
-
-/// Resolved injection region with all necessary context for LSP bridge requests
-pub struct ResolvedInjection {
-    /// Cacheable injection region with line range information
-    pub region: CacheableInjectionRegion,
-    /// Language of the injection content
-    pub injection_language: String,
-    /// Extracted virtual document content (clean, with blockquote prefixes stripped)
-    pub virtual_content: String,
-    /// Per-virtual-line column offsets for coordinate translation.
-    /// Each entry is the UTF-16 column offset for that virtual line.
-    pub line_column_offsets: Vec<u32>,
-}
-
-/// Central service for resolving injection regions at LSP positions
-pub struct InjectionResolver;
-
-impl InjectionResolver {
-    /// Resolve the injection region (if any) covering `byte_offset`. Shared by
-    /// LSP handlers (hover, completion, definition, …) at a cursor position.
-    ///
-    /// Does not hold any Document lock: inputs (`tree`, `text`) must be pre-cloned,
-    /// typically via `DocumentSnapshot`.
-    pub(crate) fn resolve_at_byte_offset(
-        coordinator: &LanguageCoordinator,
-        tracker: &NodeTracker,
-        uri: &Url,
-        tree: &Tree,
-        text: &str,
-        injection_query: &Query,
-        byte_offset: usize,
-    ) -> Option<ResolvedInjection> {
-        let injections = collect_all_injections(&tree.root_node(), text, Some(injection_query))?;
-        let (_region_index, region) = find_injection_at_position(&injections, byte_offset)?;
-        Some(Self::build_resolved_injection(
-            coordinator,
-            tracker,
-            uri,
-            region,
-            text,
-        ))
-    }
-
-    /// Calculate a stable ULID-based region_id for an injection.
-    ///
-    /// Phase 2 (lazy-node-identity-tracking): keyed on position (start_byte,
-    /// end_byte, kind, layer), so the ULID stays constant for the same position
-    /// key. Region IDs always use `layer = 0` (see below).
-    ///
-    /// Minted via the host-layer `get_or_create` (`layer = 0`) **by design**:
-    /// `content_node` comes from the host document's injection query, so it is a
-    /// host-tree node. The `layer` discriminator (lazy-node-identity-tracking
-    /// §"Node Uniqueness Key") only needs to separate host from injected nodes,
-    /// and a region's content node is always the host side. If a navigation
-    /// `kakehashi/node` call resolves that same host node it dedups to this same
-    /// ULID, which is correct.
-    pub(crate) fn calculate_region_id(
-        tracker: &NodeTracker,
-        uri: &Url,
-        injection: &InjectionRegionInfo,
-    ) -> Ulid {
-        tracker.get_or_create(
-            uri,
-            injection.content_node.start_byte(),
-            injection.content_node.end_byte(),
-            injection.content_node.kind(),
-        )
-    }
-
-    /// Resolve injection language using the unified detection chain (language-detection-fallback-chain).
-    ///
-    /// This normalizes raw fence identifiers (e.g., "py") to canonical language names
-    /// (e.g., "python") that match bridge server configurations.
-    ///
-    /// Falls back to the raw identifier if no resolution is found, allowing the
-    /// bridge lookup to fail gracefully with a clear error message.
-    fn resolve_language(
-        coordinator: &LanguageCoordinator,
-        raw_identifier: &str,
-        content: &str,
-    ) -> String {
-        coordinator
-            .detect_language(
-                raw_identifier, // Use identifier as path for token extraction
-                content,
-                Some(raw_identifier), // Explicit token
-                Some(raw_identifier), // Also try as languageId
-            )
-            .unwrap_or_else(|| raw_identifier.to_string())
-    }
-
-    /// Build a [`ResolvedInjection`] from a raw injection region.
-    ///
-    /// Shared by [`Self::resolve_at_byte_offset`] and [`Self::resolve_all`] to
-    /// avoid duplicating the region-id → cacheable-region → virtual-content →
-    /// language-resolution pipeline.
-    fn build_resolved_injection(
-        coordinator: &LanguageCoordinator,
-        tracker: &NodeTracker,
-        uri: &Url,
-        region: &InjectionRegionInfo,
-        text: &str,
-    ) -> ResolvedInjection {
-        let region_id = Self::calculate_region_id(tracker, uri, region);
-        let region_id_str = region_id.to_string();
-        let cacheable_region =
-            CacheableInjectionRegion::from_region_info(region, &region_id_str, text);
-        let (virtual_content, line_column_offsets) =
-            extract_virtual_content_and_offsets(region, &cacheable_region, text);
-        let resolved_language =
-            Self::resolve_language(coordinator, &region.language, &virtual_content);
-        ResolvedInjection {
-            region: cacheable_region,
-            injection_language: resolved_language,
-            virtual_content,
-            line_column_offsets,
-        }
-    }
-
-    /// Resolve every injection region in the document (whole-doc operations
-    /// like `documentLink` use this rather than a position lookup). Holds no
-    /// Document lock — `tree`/`text` must already be cloned, typically via
-    /// `DocumentSnapshot`. Empty vec when nothing matches.
-    pub(crate) fn resolve_all(
-        coordinator: &LanguageCoordinator,
-        tracker: &NodeTracker,
-        uri: &Url,
-        tree: &Tree,
-        text: &str,
-        injection_query: &Query,
-    ) -> Vec<ResolvedInjection> {
-        let Some(injections) =
-            collect_all_injections(&tree.root_node(), text, Some(injection_query))
-        else {
-            return Vec::new();
-        };
-
-        injections
-            .iter()
-            .map(|region| Self::build_resolved_injection(coordinator, tracker, uri, region, text))
-            .collect()
-    }
-}
+pub use discovery::{
+    CacheableInjectionRegion, InjectionRegionInfo, InjectionResolver, ResolvedInjection,
+    collect_all_injections,
+};
+pub use offset::InjectionOffset;
+
+pub(crate) use content::{
+    byte_to_point, byte_to_point_anchored, extract_clean_content, parse_with_ranges,
+};
+pub(crate) use discovery::detect_injection;
+pub(crate) use offset::effective_offset_for_pattern;
+pub(crate) use ranges::{
+    compute_included_ranges, compute_included_ranges_clipped, has_combined_for_pattern,
+    has_include_children_for_pattern, intersect_included_ranges, sub_select_included_ranges,
+};
+
+// Exposed only for the unit tests below and the selection-range tests, which
+// exercise these helpers directly. Gating with cfg(test) keeps them out of the
+// production surface so non-test builds don't see unused re-exports.
+#[cfg(test)]
+pub(crate) use content::compute_line_column_offsets;
+#[cfg(test)]
+pub(crate) use discovery::{find_injection_at_position, is_node_within};
+#[cfg(test)]
+pub(crate) use offset::parse_offset_directive_for_pattern;
 
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::language::LanguageCoordinator;
     use rstest::rstest;
-    use tree_sitter::Parser;
+    use tree_sitter::{Node, Parser, Query, StreamingIterator};
+    use url::Url;
 
     #[test]
     fn test_parse_offset_directive_for_pattern() {
