+++ src/language/injection/discovery.rs
@@ -1,0 +1,562 @@
+//! Discovering injection regions in a parse tree and resolving them into
+//! cacheable, virtual-document-ready form for the LSP bridge.
+
+use std::ops::Range;
+
+use tree_sitter::{Node, Query, QueryCapture, QueryCursor, QueryMatch, StreamingIterator, Tree};
+use ulid::Ulid;
+use url::Url;
+
+use super::content::{compute_line_column_offsets, extract_clean_content};
+use super::language::extract_injection_language;
+use super::offset::{InjectionOffset, effective_offset_for_pattern};
+use super::ranges::{
+    compute_included_ranges, compute_included_ranges_clipped, has_include_children_for_pattern,
+};
+use crate::language::LanguageCoordinator;
+use crate::language::node_tracker::NodeTracker;
+use crate::language::query_predicates::check_predicate;
+use crate::text::fnv1a_hash;
+
+/// Iterates over `@injection.content` captures that pass general predicate filtering.
+///
+/// Combines predicate evaluation (e.g., `#lua-match?`) and capture name checking
+/// into a single iterator, eliminating duplication between `collect_all_injections`
+/// and `extract_content_and_language`.
+fn iter_valid_injection_content_captures<'a, 'b>(
+    match_: &'b QueryMatch<'_, 'a>,
+    query: &'b Query,
+    text: &'b str,
+) -> impl Iterator<Item = QueryCapture<'a>> + 'b {
+    match_.captures.iter().copied().filter(move |capture| {
+        if !check_predicate(query, match_, capture, text) {
+            return false;
+        }
+        query
+            .capture_names()
+            .get(capture.index as usize)
+            .is_some_and(|name| *name == "injection.content")
+    })
+}
+
+/// Checks if a node is within the bounds of another node
+pub(crate) fn is_node_within(node: &Node, container: &Node) -> bool {
+    node.start_byte() >= container.start_byte() && node.end_byte() <= container.end_byte()
+}
+
+/// Represents an injection region found in the document
+#[derive(Debug, Clone)]
+pub struct InjectionRegionInfo<'a> {
+    /// The injection language (e.g., "lua", "yaml")
+    pub language: String,
+    /// The content node from the injection query
+    pub content_node: Node<'a>,
+    /// The pattern index (for offset directive lookups)
+    pub pattern_index: usize,
+    /// Whether `#set! injection.include-children` is set for this pattern.
+    /// When true, the injection parser sees the full content node (including named children).
+    /// When false, named children (e.g., `block_continuation`) should be excluded.
+    pub include_children: bool,
+    /// The `#offset!` directive for this pattern, resolved at collection time
+    /// (the query goes out of scope before consumers like
+    /// `CacheableInjectionRegion::from_region_info` run). `None` when the
+    /// pattern has no directive.
+    pub offset: Option<InjectionOffset>,
+}
+
+/// Owned injection region for caching (no lifetime dependency on parse tree)
+///
+/// Unlike `InjectionRegionInfo<'a>`, this struct owns all its data and can be
+/// stored in caches that outlive the parse tree. Created via `from_region_info()`.
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct CacheableInjectionRegion {
+    /// The injection language (e.g., "lua", "yaml")
+    pub language: String,
+    /// Byte range of the injection content in the source
+    pub byte_range: Range<usize>,
+    /// Line range (0-indexed, inclusive start, exclusive end)
+    pub line_range: Range<u32>,
+    /// Column offset of the injection content start within its start line,
+    /// stored as UTF-16 code units (matching LSP's position encoding).
+    /// Non-zero for inline injections (e.g., code within a markdown paragraph).
+    /// Only affects the first line (line 0) of the virtual document during
+    /// coordinate translation; subsequent lines start at column 0.
+    pub start_column: u32,
+    /// Unique identifier for associating with cached tokens.
+    /// Generated by NodeTracker as a position-based ULID.
+    pub region_id: String,
+    /// Hash of the injection content for cache invalidation.
+    /// Used to detect when cached semantic tokens should be invalidated:
+    /// if content_hash changes (or language changes), the cached tokens are stale.
+    pub content_hash: u64,
+}
+
+/// Compute the (row, UTF-16 column) of `byte_pos`, using a nearby byte with a
+/// known row as an anchor: the row counts newlines only in the span between
+/// them (offset deltas are typically a few lines, never the whole document).
+/// The column additionally scans back from `byte_pos` to its line start —
+/// bounded by the current line's length, or by `byte_pos` itself in a
+/// document without newlines.
+fn position_of_byte(
+    text: &str,
+    byte_pos: usize,
+    anchor_byte: usize,
+    anchor_row: usize,
+) -> (u32, u32) {
+    let (lo, hi) = (byte_pos.min(anchor_byte), byte_pos.max(anchor_byte));
+    let newlines = text[lo..hi].bytes().filter(|&b| b == b'\n').count();
+    let row = if byte_pos >= anchor_byte {
+        anchor_row + newlines
+    } else {
+        // Defensive: a stale tree whose row metadata disagrees with `text`
+        // must not panic the server on underflow.
+        anchor_row.saturating_sub(newlines)
+    };
+    let line_start = text[..byte_pos].rfind('\n').map(|i| i + 1).unwrap_or(0);
+    let column = text[line_start..byte_pos].encode_utf16().count() as u32;
+    (row as u32, column)
+}
+
+/// Compute clean virtual content and per-line column offsets for an injection region.
+///
+/// This combines `compute_included_ranges`, `extract_clean_content`, and
+/// `compute_line_column_offsets` into a single call, avoiding duplication
+/// across resolve methods.
+fn extract_virtual_content_and_offsets(
+    region: &InjectionRegionInfo,
+    cacheable: &CacheableInjectionRegion,
+    text: &str,
+) -> (String, Vec<u32>) {
+    // Child-exclusion gaps restricted to the effective window (#186):
+    // cacheable.byte_range already reflects any #offset! directive (applied
+    // and char-boundary-aligned by from_region_info), so gaps are clipped to
+    // it and relativized to its start. Without an offset the window equals
+    // the content node span, so the node-anchored variant gives the same
+    // result while reusing tree-sitter's cached node Points (no text scan).
+    let included_ranges = if region.offset.is_some() {
+        compute_included_ranges_clipped(
+            &region.content_node,
+            region.include_children,
+            text,
+            cacheable.byte_range.clone(),
+        )
+    } else {
+        compute_included_ranges(&region.content_node, region.include_children)
+    };
+    let virtual_content = extract_clean_content(
+        text,
+        cacheable.byte_range.clone(),
+        included_ranges.as_deref(),
+    );
+    let line_column_offsets = compute_line_column_offsets(
+        text,
+        cacheable.byte_range.clone(),
+        cacheable.start_column,
+        included_ranges.as_deref(),
+    );
+    (virtual_content, line_column_offsets)
+}
+
+impl CacheableInjectionRegion {
+    /// Create from an InjectionRegionInfo, extracting position data from the node.
+    ///
+    /// # Panics (debug builds)
+    /// Panics if the content node is zero-width (`start_byte >= end_byte`).
+    /// Callers must filter zero-width nodes upstream (see `collect_all_injections`).
+    pub fn from_region_info(info: &InjectionRegionInfo<'_>, region_id: &str, text: &str) -> Self {
+        let node = &info.content_node;
+        debug_assert!(
+            node.start_byte() < node.end_byte(),
+            "from_region_info called with zero-width node at byte {}",
+            node.start_byte(),
+        );
+
+        // #offset! narrows the raw node span to the effective content range
+        // (e.g. trimming `---` frontmatter delimiters). The bridge consumes
+        // byte_range / line_range / start_column for virtual-document content
+        // extraction and coordinate translation, so all of them must reflect
+        // the effective range, mirroring the semantic path (#183).
+        let (start_byte, end_byte) = match info.offset {
+            Some(offset) => {
+                use crate::analysis::offset_calculator::{ByteRange, calculate_effective_range};
+                // calculate_effective_range clamps, snaps inward to char
+                // boundaries, and normalizes start <= end, so slicing below
+                // cannot panic.
+                let effective = calculate_effective_range(
+                    text,
+                    ByteRange::new(node.start_byte(), node.end_byte()),
+                    offset,
+                );
+                (effective.start, effective.end)
+            }
+            None => (node.start_byte(), node.end_byte()),
+        };
+
+        let content = &text[start_byte..end_byte];
+
+        // Convert tree-sitter byte column to UTF-16 code units for LSP compatibility.
+        // Tree-sitter reports columns as byte offsets, but LSP positions use UTF-16.
+        // For ASCII they're identical, but non-ASCII prefixes (CJK, emoji) differ.
+        // tree-sitter's `column` is a byte offset from the line start, so
+        // `start_byte() - column` correctly recovers the line-start byte even
+        // when the line contains multi-byte UTF-8 characters before the node.
+        let (start_line, start_column) = if start_byte == node.start_byte() {
+            let byte_column = node.start_position().column;
+            let line_start_byte = node.start_byte() - byte_column;
+            let line_prefix = &text[line_start_byte..node.start_byte()];
+            (
+                node.start_position().row as u32,
+                line_prefix.encode_utf16().count() as u32,
+            )
+        } else {
+            position_of_byte(
+                text,
+                start_byte,
+                node.start_byte(),
+                node.start_position().row,
+            )
+        };
+
+        // end_position() points past the last byte. When that position is mid-line
+        // (column > 0), the node still occupies that row → exclusive end is row + 1.
+        // When column == 0 (node ended with newline), the row is already one past
+        // the last occupied line — use it directly.
+        let end_line = if end_byte == node.end_byte() {
+            let end_pos = node.end_position();
+            if end_pos.column > 0 {
+                end_pos.row as u32 + 1
+            } else {
+                end_pos.row as u32
+            }
+        } else {
+            let (end_row, end_column) =
+                position_of_byte(text, end_byte, node.end_byte(), node.end_position().row);
+            if end_column > 0 { end_row + 1 } else { end_row }
+        };
+
+        Self {
+            language: info.language.clone(),
+            byte_range: start_byte..end_byte,
+            line_range: start_line..end_line,
+            start_column,
+            region_id: region_id.to_string(),
+            content_hash: Self::hash_content(content),
+        }
+    }
+
+    /// Compute a simple hash of content bytes for stable matching.
+    fn hash_content(content: &str) -> u64 {
+        fnv1a_hash(content)
+    }
+}
+
+/// Collects all injection regions in the document.
+///
+/// Unlike `detect_injection` (which requires a specific node), this finds ALL
+/// injection regions in the whole document — used by semantic tokens to highlight
+/// every injected region. `None` when there is no query.
+pub fn collect_all_injections<'a>(
+    root: &Node<'a>,
+    text: &str,
+    injection_query: Option<&Query>,
+) -> Option<Vec<InjectionRegionInfo<'a>>> {
+    let query = injection_query?;
+
+    let mut cursor = QueryCursor::new();
+    let mut matches = cursor.matches(query, *root, text.as_bytes());
+
+    // Use a map to deduplicate by content node range
+    let mut injections_map = std::collections::HashMap::new();
+
+    while let Some(match_) = matches.next() {
+        for capture in iter_valid_injection_content_captures(match_, query, text) {
+            if capture.node.start_byte() >= capture.node.end_byte() {
+                continue;
+            }
+            if let Some(language) = extract_injection_language(query, match_, text) {
+                let key = (capture.node.start_byte(), capture.node.end_byte());
+                injections_map.entry(key).or_insert(InjectionRegionInfo {
+                    language,
+                    content_node: capture.node,
+                    pattern_index: match_.pattern_index,
+                    include_children: has_include_children_for_pattern(query, match_.pattern_index),
+                    offset: effective_offset_for_pattern(query, match_.pattern_index),
+                });
+            }
+        }
+    }
+
+    // Sort by start_byte (primary) and end_byte (secondary) to ensure deterministic ordering
+    let mut injections: Vec<_> = injections_map.into_values().collect();
+    injections.sort_by_key(|r| (r.content_node.start_byte(), r.content_node.end_byte()));
+    Some(injections)
+}
+
+/// Detects injection and returns both the language and the content node
+/// Also returns the pattern index of the innermost injection for offset lookups
+pub(crate) fn detect_injection<'a>(
+    node: &Node<'a>,
+    root: &Node<'a>,
+    text: &str,
+    injection_query: Option<&Query>,
+    base_language: &str,
+) -> Option<(Vec<String>, Node<'a>, usize)> {
+    let injections = collect_injection_regions(node, root, text, injection_query)?;
+
+    if injections.is_empty() {
+        return None;
+    }
+
+    // Sort injections by their range (outer to inner)
+    let mut sorted_injections = injections;
+    sorted_injections.sort_by(|a, b| {
+        // Sort by start byte (ascending), then by end byte (descending)
+        // This ensures outer injections come before inner ones
+        a.start_byte
+            .cmp(&b.start_byte)
+            .then(b.end_byte.cmp(&a.end_byte))
+    });
+
+    // Build the language hierarchy from outermost to innermost
+    let mut hierarchy = vec![base_language.to_string()];
+    for region in &sorted_injections {
+        hierarchy.push(region.language.clone());
+    }
+
+    // Return the innermost content node and its pattern index
+    let innermost = sorted_injections.last()?;
+
+    Some((hierarchy, innermost.content_node, innermost.pattern_index))
+}
+
+/// Represents an injection region with its metadata (unresolved intermediate data).
+struct RawInjectionRegion<'a> {
+    start_byte: usize,
+    end_byte: usize,
+    language: String,
+    content_node: Node<'a>,
+    pattern_index: usize,
+}
+
+/// Collects all injection regions that contain the given node
+/// Returns a list of `RawInjectionRegion` values for each injection containing the node.
+fn collect_injection_regions<'a>(
+    node: &Node<'a>,
+    root: &Node<'a>,
+    text: &str,
+    injection_query: Option<&Query>,
+) -> Option<Vec<RawInjectionRegion<'a>>> {
+    let query = injection_query?;
+
+    // Run the query on the entire tree
+    let mut cursor = QueryCursor::new();
+    let mut matches = cursor.matches(query, *root, text.as_bytes());
+
+    // Collect all injection regions that contain our node
+    // Use a map to deduplicate by node range (start, end)
+    let mut injections_map = std::collections::HashMap::new();
+
+    while let Some(match_) = matches.next() {
+        if let Some((content_node, language, pattern_index)) =
+            extract_content_and_language(node, match_, query, text)
+        {
+            let key = (content_node.start_byte(), content_node.end_byte());
+
+            // Only keep the first injection for each unique range
+            // This handles cases where multiple patterns match the same node
+            injections_map.entry(key).or_insert(RawInjectionRegion {
+                start_byte: content_node.start_byte(),
+                end_byte: content_node.end_byte(),
+                language,
+                content_node,
+                pattern_index,
+            });
+        }
+    }
+
+    // Convert to vector
+    let injections: Vec<_> = injections_map.into_values().collect();
+
+    Some(injections)
+}
+
+/// Extracts the injection content node and language if the given node is within it
+/// Also returns the pattern index for offset lookups
+fn extract_content_and_language<'a>(
+    node: &Node<'a>,
+    match_: &QueryMatch<'_, 'a>,
+    query: &Query,
+    text: &str,
+) -> Option<(Node<'a>, String, usize)> {
+    for capture in iter_valid_injection_content_captures(match_, query, text) {
+        let content_node = capture.node;
+
+        // Check if our node is within this injection region
+        if is_node_within(node, &content_node) {
+            // Extract the injection language
+            if let Some(language) = extract_injection_language(query, match_, text) {
+                return Some((content_node, language, match_.pattern_index));
+            }
+        }
+    }
+
+    None
+}
+
+/// Find the injection region containing the given byte offset.
+///
+/// Returns `(index, region)` for use with `calculate_region_id`, or `None` when the
+/// position is not within any injection region.
+pub(crate) fn find_injection_at_position<'a>(
+    injections: &'a [InjectionRegionInfo<'a>],
+    byte_offset: usize,
+) -> Option<(usize, &'a InjectionRegionInfo<'a>)> {
+    injections.iter().enumerate().find(|(_, inj)| {
+        let start = inj.content_node.start_byte();
+        let end = inj.content_node.end_byte();
+        byte_offset >= start && byte_offset < end
+    })
+}
+
+/// Resolved injection region with all necessary context for LSP bridge requests
+pub struct ResolvedInjection {
+    /// Cacheable injection region with line range information
+    pub region: CacheableInjectionRegion,
+    /// Language of the injection content
+    pub injection_language: String,
+    /// Extracted virtual document content (clean, with blockquote prefixes stripped)
+    pub virtual_content: String,
+    /// Per-virtual-line column offsets for coordinate translation.
+    /// Each entry is the UTF-16 column offset for that virtual line.
+    pub line_column_offsets: Vec<u32>,
+}
+
+/// Central service for resolving injection regions at LSP positions
+pub struct InjectionResolver;
+
+impl InjectionResolver {
+    /// Resolve the injection region (if any) covering `byte_offset`. Shared by
+    /// LSP handlers (hover, completion, definition, …) at a cursor position.
+    ///
+    /// Does not hold any Document lock: inputs (`tree`, `text`) must be pre-cloned,
+    /// typically via `DocumentSnapshot`.
+    pub(crate) fn resolve_at_byte_offset(
+        coordinator: &LanguageCoordinator,
+        tracker: &NodeTracker,
+        uri: &Url,
+        tree: &Tree,
+        text: &str,
+        injection_query: &Query,
+        byte_offset: usize,
+    ) -> Option<ResolvedInjection> {
+        let injections = collect_all_injections(&tree.root_node(), text, Some(injection_query))?;
+        let (_region_index, region) = find_injection_at_position(&injections, byte_offset)?;
+        Some(Self::build_resolved_injection(
+            coordinator,
+            tracker,
+            uri,
+            region,
+            text,
+        ))
+    }
+
+    /// Calculate a stable ULID-based region_id for an injection.
+    ///
+    /// Phase 2 (lazy-node-identity-tracking): keyed on position (start_byte,
+    /// end_byte, kind, layer), so the ULID stays constant for the same position
+    /// key. Region IDs always use `layer = 0` (see below).
+    ///
+    /// Minted via the host-layer `get_or_create` (`layer = 0`) **by design**:
+    /// `content_node` comes from the host document's injection query, so it is a
+    /// host-tree node. The `layer` discriminator (lazy-node-identity-tracking
+    /// §"Node Uniqueness Key") only needs to separate host from injected nodes,
+    /// and a region's content node is always the host side. If a navigation
+    /// `kakehashi/node` call resolves that same host node it dedups to this same
+    /// ULID, which is correct.
+    pub(crate) fn calculate_region_id(
+        tracker: &NodeTracker,
+        uri: &Url,
+        injection: &InjectionRegionInfo,
+    ) -> Ulid {
+        tracker.get_or_create(
+            uri,
+            injection.content_node.start_byte(),
+            injection.content_node.end_byte(),
+            injection.content_node.kind(),
+        )
+    }
+
+    /// Resolve injection language using the unified detection chain (language-detection-fallback-chain).
+    ///
+    /// This normalizes raw fence identifiers (e.g., "py") to canonical language names
+    /// (e.g., "python") that match bridge server configurations.
+    ///
+    /// Falls back to the raw identifier if no resolution is found, allowing the
+    /// bridge lookup to fail gracefully with a clear error message.
+    fn resolve_language(
+        coordinator: &LanguageCoordinator,
+        raw_identifier: &str,
+        content: &str,
+    ) -> String {
+        coordinator
+            .detect_language(
+                raw_identifier, // Use identifier as path for token extraction
+                content,
+                Some(raw_identifier), // Explicit token
+                Some(raw_identifier), // Also try as languageId
+            )
+            .unwrap_or_else(|| raw_identifier.to_string())
+    }
+
+    /// Build a [`ResolvedInjection`] from a raw injection region.
+    ///
+    /// Shared by [`Self::resolve_at_byte_offset`] and [`Self::resolve_all`] to
+    /// avoid duplicating the region-id → cacheable-region → virtual-content →
+    /// language-resolution pipeline.
+    fn build_resolved_injection(
+        coordinator: &LanguageCoordinator,
+        tracker: &NodeTracker,
+        uri: &Url,
+        region: &InjectionRegionInfo,
+        text: &str,
+    ) -> ResolvedInjection {
+        let region_id = Self::calculate_region_id(tracker, uri, region);
+        let region_id_str = region_id.to_string();
+        let cacheable_region =
+            CacheableInjectionRegion::from_region_info(region, &region_id_str, text);
+        let (virtual_content, line_column_offsets) =
+            extract_virtual_content_and_offsets(region, &cacheable_region, text);
+        let resolved_language =
+            Self::resolve_language(coordinator, &region.language, &virtual_content);
+        ResolvedInjection {
+            region: cacheable_region,
+            injection_language: resolved_language,
+            virtual_content,
+            line_column_offsets,
+        }
+    }
+
+    /// Resolve every injection region in the document (whole-doc operations
+    /// like `documentLink` use this rather than a position lookup). Holds no
+    /// Document lock — `tree`/`text` must already be cloned, typically via
+    /// `DocumentSnapshot`. Empty vec when nothing matches.
+    pub(crate) fn resolve_all(
+        coordinator: &LanguageCoordinator,
+        tracker: &NodeTracker,
+        uri: &Url,
+        tree: &Tree,
+        text: &str,
+        injection_query: &Query,
+    ) -> Vec<ResolvedInjection> {
+        let Some(injections) =
+            collect_all_injections(&tree.root_node(), text, Some(injection_query))
+        else {
+            return Vec::new();
+        };
+
+        injections
+            .iter()
+            .map(|region| Self::build_resolved_injection(coordinator, tracker, uri, region, text))
+            .collect()
+    }
+}
